passage de long double à double

+++ NN.h
@@ -2,10 +2,10 @@
 
 /********************************************************
 Cette fonction réalise l'opération de convulation sur
-la matrice contenant les valeurs des pixel de l'image 
+la matrice contenant les valeurs des pixel de l'image
  avec la matrice de poid qui lui est associée
 ************************************************************/
-void convolution(long double *W_C, long double *CONVOLUTION, long double *IMAGE){
+void convolution(double *W_C, double *CONVOLUTION, double *IMAGE){
 	for(int i=0; i<(TAILLE_IMAGE-2); i++){
 		for (int j=0; j<(TAILLE_IMAGE-2); j++){
 			int offset2 = i*(TAILLE_IMAGE-2)+j;
@@ -19,7 +19,7 @@ void convolution(long double *W_C, long double *CONVOLUTION, long double *IMAGE)
 	}
 }
 
-void pooling(long double *CONVOLUTION, long double *POOLING){
+void pooling(double *CONVOLUTION, double *POOLING){
 	int compare, compare2;
 	for(int k=0; k<TAILLE_POOL; k++){
 		for (int i=0; i<TAILLE_IMAGE-2; i+=2){
@@ -49,7 +49,7 @@ void pooling(long double *CONVOLUTION, long double *POOLING){
 	}
 }
 
-void convolution2(long double *W_C2, long double *CONVOLUTION2, long double *POOLING){
+void convolution2(double *W_C2, double *CONVOLUTION2, double *POOLING){
 	for(int i=0; i<(DIM_POOL-2); i++){
 		for (int j=0; j<(DIM_POOL-2); j++){
 			int offset2 = i*(DIM_POOL-2)+j;
@@ -63,7 +63,7 @@ void convolution2(long double *W_C2, long double *CONVOLUTION2, long double *POO
 	}
 }
 
-void pooling2(long double *CONVOLUTION2, long double *POOLING2){
+void pooling2(double *CONVOLUTION2, double *POOLING2){
 	int compare, compare2;
 	for(int k=0; k<TAILLE_POOLII; k++){
 		for (int i=0; i<DIM_POOL-2; i+=2){
@@ -98,12 +98,12 @@ void pooling2(long double *CONVOLUTION2, long double *POOLING2){
 	de l'output à partir des données contenu dans l'input
 **************************************************************************/
 
-void front_prop(long double *Z_L,long double *Z_lIII,long double *Z_lII,
-								long double *Z_l, long double *INPUT,long double *LAYER,
-								long double *LAYERII, long double *LAYERIII, long double *OUTPUT,
-								long double *b_L, long double *b_LII, long double *b_LIII,
-								long double *b_output,long double* W_L, long double *W_LII,
-								long double *W_LIII, long double* W_O){
+void front_prop(double *Z_L,double *Z_lIII,double *Z_lII,
+								double *Z_l, double *INPUT,double *LAYER,
+								double *LAYERII, double *LAYERIII, double *OUTPUT,
+								double *b_L, double *b_LII, double *b_LIII,
+								double *b_output,double* W_L, double *W_LII,
+								double *W_LIII, double* W_O){
 		//premiere propagation Input à Layer
 	//	#pragma omp parallel shared(Z_L, Z_lIII, Z_lII, Z_l,INPUT,LAYER,LAYERII,LAYERIII,OUTPUT,b_L,b_LII, b_LIII,b_output,W_L,W_LII, W_LIII, W_O)
 		//{
@@ -154,8 +154,8 @@ error_output calcul l'erreur sur l'output
 à partir de la solution
 ************************************************/
 
-void error_output(long double *Z_L,long double *ERROR_OUTPUT,
-                  long double *SOLUTION,long double *OUTPUT){
+void error_output(double *Z_L,double *ERROR_OUTPUT,
+                  double *SOLUTION,double *OUTPUT){
   for (int j=0; j<TAILLE_OUTPUT; j = j+1){
     ERROR_OUTPUT[j] = (OUTPUT[j]-SOLUTION[j])*d_sigmoid(Z_L[j]);
   }
@@ -167,8 +167,8 @@ calcul de l'erreur aux niveaux du
 l'output et par une multiplication
 de la transpose de sa matrice des poids
 **********************************************/
-void error_layerIII(long double *ans3, long double *W_OT, long double *ERROR_LAYERIII,
-                  	long double *ERROR_OUTPUT, long double *Z_lIII){
+void error_layerIII(double *ans3, double *W_OT, double *ERROR_LAYERIII,
+                  	double *ERROR_OUTPUT, double *Z_lIII){
   multiplAV(W_OT, ERROR_OUTPUT, ans3, TAILLE_LAYERIII, TAILLE_OUTPUT);
   for (int j=0; j<TAILLE_LAYERIII; j = j+1){
     ERROR_LAYERIII[j] = ans3[j]*d_sigmoid(Z_lIII[j]);
@@ -182,8 +182,8 @@ le 3ieme layers et par une multiplication
 de la transpose de sa matrice des poids
 **********************************************/
 
-void error_layerII(long double *ans2, long double *W_LIIIT, long double *ERROR_LAYERII,
-									 long double *ERROR_LAYERIII, long double *Z_lII){
+void error_layerII(double *ans2, double *W_LIIIT, double *ERROR_LAYERII,
+									 double *ERROR_LAYERIII, double *Z_lII){
 	multiplAV(W_LIIIT, ERROR_LAYERIII, ans2, TAILLE_LAYERII, TAILLE_LAYERIII);
 	#pragma omp for nowait
 	for (int j=0; j<TAILLE_LAYERII; j++){
@@ -197,8 +197,8 @@ calcul de l'erreur aux niveaux du
 le 2nd layer et par une multiplication
 de la transpose de sa matrice  des poids
 **********************************************/
-void error_layer(long double *ans, long double *W_LIIT, long double *ERROR_LAYER,
- 								 long double *ERROR_LAYERII, long double *Z_l){
+void error_layer(double *ans, double *W_LIIT, double *ERROR_LAYER,
+ 								 double *ERROR_LAYERII, double *Z_l){
 	multiplAV(W_LIIT, ERROR_LAYERII, ans, TAILLE_LAYER, TAILLE_LAYERII);
 	#pragma omp for nowait
 	for (int j=0; j< TAILLE_LAYER; j++){
@@ -210,13 +210,13 @@ void error_layer(long double *ans, long double *W_LIIT, long double *ERROR_LAYER
  backprop() la fonction qui réalise la back-propagation
  et qui corrige les valeurs de poids et de biais a chaque apprentissage
  **********************************************************************/
-void backprop(long double *W_L, long double *W_LII, long double *W_LIII,
-	 						long double *W_O, long double *b_L, long double *b_LII,
-							long double *b_LIII, long double *b_O, long double *eta,
-							long double *ERROR_OUTPUT, long double *ERROR_LAYERIII,
-							long double *ERROR_LAYERII, long double *ERROR_LAYER,long double *INPUT,
-							long double *LAYER, long double *LAYERII, long double *LAYERIII,
-							long double *OUTPUT){
+void backprop(double *W_L, double *W_LII, double *W_LIII,
+	 						double *W_O, double *b_L, double *b_LII,
+							double *b_LIII, double *b_O, double *eta,
+							double *ERROR_OUTPUT, double *ERROR_LAYERIII,
+							double *ERROR_LAYERII, double *ERROR_LAYER,double *INPUT,
+							double *LAYER, double *LAYERII, double *LAYERIII,
+							double *OUTPUT){
 
   // modification BIAIS et POIDS pour OUTPUT
   for (int j=0; j<TAILLE_OUTPUT; j = j+1){
